+++ CHANGELOG.md
@@ -13,12 +13,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - bespoke ways to calibrate data - such as finding only weights or equal weighting
 - spike detection using simple offset rule
 - stand alone method (`fetch_nmdb_data()`) for fetching and returning a dataframe of data from NMDB.eu
+- 3 methods for pressure correction (changes derivation of beta coefficient)
 
 ### Changed
 
 - changed attribute naming from `koehlie_method_form` to `koehlie_parameters` throughout
 - Refactored calibration routines back end
 - default `koehlie_parameters` changed from `Mar21_uranos_drf` to `Mar21_mcnp_drf`
+- config - pressure correction method `zreda_2010` renamed to `desilets_zreda_2003`
+- name of pressure corrections method in config file (breaking)
 
 ### Depreceated